requirements.txt
pandas
numpy
pyyaml
openpyxl
plotly>=5.0.0
streamlit>=1.32.0
